requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coverage==6.5.0
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-debug-toolbar==3.7.0
django-debug-toolbar-force==0.1.8
django-filter==22.1
django-nine==0.2.5
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
urllib3==1.26.12
